+++ src/sgl/device/shader.h
@@ -13,6 +13,7 @@
 #include <exception>
 #include <map>
 #include <set>
+#include <span>
 #include <string>
 #include <vector>